Revert changing binning from float to double

+++ Core/src/Tools/LayerArrayCreator.cpp
@@ -72,7 +72,7 @@ Acts::LayerArrayCreator::layerArray(const LayerVector& layersInput,
 
   // arbitrary binning
   case arbitrary: {
-    std::vector<double> boundaries;
+    std::vector<float> boundaries;
     // initial step
     boundaries.push_back(min);
     double                       layerValue     = 0.;
